LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A6_T1-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Expected an identifier but found '\u000A' instead
S7.3_A6_T1

PASS Expected parsing failure

TEST COMPLETE
